+++ tests/TestCase/Model/Table/VideosTableTest.php
@@ -88,15 +88,6 @@ class VideosTableTest extends TestCase
     }
 
     /**
-     * Test for `_getInfo()` method
-     * @test
-     */
-    public function testGetInfo()
-    {
-        $this->assertFalse($this->invokeMethod($this->Videos, '_getInfo', ['vlSR8Wlmpac']));
-    }
-
-    /**
      * Test for `afterDelete()` method
      * @test
      */
@@ -293,24 +284,43 @@ class VideosTableTest extends TestCase
     }
 
     /**
+     * Test for `_getRandomSpots()` method
+     * @test
+     */
+    public function testInternalGetRandomSpots()
+    {
+        $query = $this->invokeMethod($this->Videos, '_getRandomSpots');
+
+        $this->assertInstanceof('Cake\ORM\Query', $query);
+        $this->assertStringEndsWith(
+            'FROM youtube_videos Videos WHERE (Videos.active = :c0 AND Videos.is_spot = :c1 AND Videos.created <= :c2)',
+            $query->sql()
+        );
+
+        $this->assertTrue($query->valueBinder()->bindings()[':c0']['value']);
+        $this->assertTrue($query->valueBinder()->bindings()[':c1']['value']);
+        $this->assertInstanceof('Cake\I18n\Time', $query->valueBinder()->bindings()[':c2']['value']);
+    }
+
+    /**
      * Test for `getRandomSpots()` method
      * @test
      */
     public function testGetRandomSpots()
     {
-        $spots = $this->Videos->getRandomSpots();
+        $collection = $this->Videos->getRandomSpots();
+        $this->assertInstanceof('Cake\Collection\Collection', $collection);
 
-        $this->assertCount(1, $spots);
-        $this->assertInstanceOf('MeCmsYoutube\Model\Entity\Video', $spots[0]);
-        $this->assertNotEmpty($spots[0]->youtube_id);
+        $this->assertEquals(1, count($collection->toArray()));
+        $this->assertInstanceOf('MeCmsYoutube\Model\Entity\Video', $collection->first());
 
-        $spots = $this->Videos->getRandomSpots(2);
+        $collection = $this->Videos->getRandomSpots(2);
+        $this->assertInstanceof('Cake\Collection\Collection', $collection);
 
-        $this->assertCount(2, $spots);
+        $this->assertEquals(2, count($collection->toArray()));
 
-        foreach ($spots as $spot) {
-            $this->assertInstanceOf('MeCmsYoutube\Model\Entity\Video', $spot);
-            $this->assertNotEmpty($spot->youtube_id);
+        foreach ($collection->toArray() as $video) {
+            $this->assertInstanceOf('MeCmsYoutube\Model\Entity\Video', $video);
         }
     }