+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # 2.x branch
+## 2.10 branch
+### 2.10.0
+* `VideosTable::getRandomSpots()` now returns a `Collection` instance;
+* updated for MeCms 2.18.0;
+* added tests for all admin classes.
+
 ## 2.9 branch
+### 2.9.3
+* widgets (cells) use collections;
+* tests that require a network connection have been marked with the
+    `requireNetwork` group;
+* added tests for `beforeFilter()` and `initialize()` methods of all admin
+    controllers.
+
 ### 2.9.2
 * fixed `VideosController::indexByDate()` method;
 * fixed the `findActive()` method for `VideosCategoriesTable`;